Module kanidm_proto::v1

Structs

AccountUnixExtend

ApiToken

ApiTokenGenerate

AuthRequest

AuthResponse

BackupCodesView

CUIntentToken

CUSessionToken

CUStatus

Claim

CreateRequest

CredentialDetail

CredentialStatus

DeleteRequest

Entry

Group

GroupUnixExtend

ModifyList

ModifyRequest

PasskeyDetail

RadiusAuthToken

SearchRequest

SearchResponse

SingleStringRequest

TotpSecret

UatStatus

UnixGroupToken

UnixUserToken

UserAuthToken

The currently authenticated user, and any required metadata for them to properly authorise them. This is similar in nature to oauth and the krb PAC/PAD structures. This information is transparent to clients and CAN be parsed by them!

WhoamiResponse

Enums

ApiTokenPurpose

AuthAllowed

AuthCredential

AuthIssueSession

AuthMech

AuthState

AuthStep

AuthType

CURegState

CURequest

ConsistencyError

CredentialDetailType

Filter

Modify

OperationError

PasswordFeedback

PluginError

SchemaError

TotpAlgo

UatPurpose

UatPurposeStatus

UiHint