Module kanidm_proto::v1

Structs

AccountPersonSet

AccountUnixExtend

AuthRequest

AuthResponse

BackupCodesView

CUIntentToken

CUSessionToken

CUStatus

Claim

CreateRequest

CredentialDetail

CredentialStatus

DeleteRequest

Entry

Group

GroupUnixExtend

ModifyList

ModifyRequest

RadiusAuthToken

SearchRequest

SearchResponse

SingleStringRequest

TotpSecret

UnixGroupToken

UnixUserToken

UserAuthToken

The currently authenticated user, and any required metadata for them to properly authorise them. This is similar in nature to oauth and the krb PAC/PAD structures. Currently we only use this internally, but we should consider making it “parseable” by the client so they can have per-session group/authorisation data.

WhoamiResponse

Enums

AuthAllowed

AuthCredential

AuthMech

AuthState

AuthStep

AuthType

CURegState

CURequest

ConsistencyError

CredentialDetailType

Filter

Modify

OperationError

PasswordFeedback

PluginError

SchemaError

SetCredentialRequest

SetCredentialResponse

TotpAlgo